requirements.txt
googlesearch-python
bs4
langchain==0.0.324
openai==0.28.1